holiday stops: layout tweaks
